requirements.txt
streamlit
pandas
requests
plotly
google-generativeai
python-dotenv
Pillow
fpdf2
audio-recorder-streamlit
